fix(package): move ProcessDefinitionNamespace to avoid package collision
